Split a test into multiple smaller tests.
